# contracts/openapi/teamos.yaml
openapi: 3.1.0
...
info:
  title: TeamOS Gateway API
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: |
    Внешний JSON/REST-контракт TeamOS. Поля и формы ответов зеркалят
    `src/api/index.ts` и `src/types/index.ts` frontend-приложения.
    Все даты-время передаются в ISO-8601, идентификаторы являются UUID.
servers:
  - url: http://localhost:8080
    description: Локальный gateway
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/kb/sections:
    get:
      tags: [Knowledge Base]
      operationId: getSections
      summary: Получить разделы базы знаний
      responses:
        '200':
          $ref: '#/components/responses/ArticleSections'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Knowledge Base]
      operationId: createSection
      summary: Создать раздел базы знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateArticleSectionInput'
      responses:
        '201':
          $ref: '#/components/responses/ArticleSection'
        default:
          $ref: '#/components/responses/Error'
...
  /api/v1/academy/courses:
    get:
      tags: [Academy]
      operationId: getCourses
      summary: Получить курсы
      responses:
        '200':
          $ref: '#/components/responses/Courses'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Academy]
      operationId: createCourse
      summary: Создать курс
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseInput'
      responses:
        '201':
          $ref: '#/components/responses/Course'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/courses/from-kb:
    post:
      tags: [Academy]
      operationId: createCourseFromKb
      summary: Создать курс из разделов и статей базы знаний
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseFromKbInput'
      responses:
        '201':
          $ref: '#/components/responses/Course'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/courses/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    get:
      tags: [Academy]
      operationId: getCourse
      summary: Получить курс
      responses:
        '200':
          $ref: '#/components/responses/Course'
        default:
          $ref: '#/components/responses/Error'
    patch:
      tags: [Academy]
      operationId: updateCourse
      summary: Обновить курс
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseInput'
      responses:
        '200':
          $ref: '#/components/responses/Course'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Academy]
      operationId: deleteCourse
      summary: Удалить курс каскадно
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/courses/{courseId}/sections:
    parameters:
      - name: courseId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    get:
      tags: [Academy]
      operationId: getCourseSections
      summary: Получить разделы курса
      responses:
        '200':
          $ref: '#/components/responses/CourseSections'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Academy]
      operationId: createCourseSection
      summary: Создать раздел курса
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCourseSectionInput'
      responses:
        '201':
          $ref: '#/components/responses/CourseSection'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/sections/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Academy]
      operationId: updateCourseSection
      summary: Переименовать раздел курса
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCourseSectionInput'
      responses:
        '200':
          $ref: '#/components/responses/CourseSection'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Academy]
      operationId: deleteCourseSection
      summary: Удалить раздел курса с уроками
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/lessons:
    get:
      tags: [Academy]
      operationId: getLessons
      summary: Получить уроки
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          $ref: '#/components/responses/Lessons'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Academy]
      operationId: createLesson
      summary: Создать урок
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLessonInput'
      responses:
        '201':
          $ref: '#/components/responses/Lesson'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/lessons/{id}:
    parameters:
      - $ref: '#/components/parameters/Id'
    patch:
      tags: [Academy]
      operationId: updateLesson
      summary: Обновить урок
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateLessonInput'
      responses:
        '200':
          $ref: '#/components/responses/Lesson'
        default:
          $ref: '#/components/responses/Error'
    delete:
      tags: [Academy]
      operationId: deleteLesson
      summary: Удалить урок с тестом и прогрессом
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/lessons/{id}/move:
    parameters:
      - $ref: '#/components/parameters/Id'
    post:
      tags: [Academy]
      operationId: moveLesson
      summary: Переместить урок и изменить порядок
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MoveLessonInput'
      responses:
        '200':
          $ref: '#/components/responses/Lesson'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/quizzes:
    get:
      tags: [Academy]
      operationId: getQuizzes
      summary: Получить тесты
      parameters:
        - name: lessonId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          $ref: '#/components/responses/Quizzes'
        default:
          $ref: '#/components/responses/Error'
    put:
      tags: [Academy]
      operationId: upsertQuiz
      summary: Создать или обновить тест
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertQuizInput'
      responses:
        '200':
          $ref: '#/components/responses/Quiz'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/assignments:
    get:
      tags: [Academy]
      operationId: getAssignments
      summary: Получить назначения курсов
      responses:
        '200':
          $ref: '#/components/responses/CourseAssignments'
        default:
          $ref: '#/components/responses/Error'
    post:
      tags: [Academy]
      operationId: assignCourse
      summary: Назначить курс
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignCourseInput'
      responses:
        '201':
          $ref: '#/components/responses/CourseAssignment'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/progress:
    get:
      tags: [Academy]
      operationId: getProgress
      summary: Получить прогресс по курсам
      parameters:
        - name: courseId
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/ID'
      responses:
        '200':
          $ref: '#/components/responses/CourseProgressList'
        default:
          $ref: '#/components/responses/Error'
  /api/v1/academy/progress/lessons/{lessonId}/complete:
    parameters:
      - name: lessonId
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ID'
    post:
      tags: [Academy]
      operationId: markLessonComplete
      summary: Отметить урок пройденным
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkLessonCompleteInput'
      responses:
        '200':
          $ref: '#/components/responses/CourseProgress'
        default:
          $ref: '#/components/responses/Error'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    refreshCookie:
      type: apiKey
      in: cookie
      name: teamos_refresh
...
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        $ref: '#/components/schemas/ID'
...
  responses:
    NoContent:
      description: Операция выполнена
...
    Error:
      description: Ошибка API
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ArticleSection:
      description: Раздел базы знаний
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ArticleSection'
    ArticleSections:
      description: Разделы базы знаний
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/ArticleSection'
...
    Course:
      description: Курс
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Course'
    Courses:
      description: Курсы
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Course'
    CourseSection:
      description: Раздел курса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CourseSection'
    CourseSections:
      description: Разделы курса
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CourseSection'
    Lesson:
      description: Урок
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Lesson'
    Lessons:
      description: Уроки
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Lesson'
    Quiz:
      description: Тест урока
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Quiz'
    Quizzes:
      description: Тесты уроков
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/Quiz'
    CourseAssignment:
      description: Назначение курса
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CourseAssignment'
    CourseAssignments:
      description: Назначения курсов
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CourseAssignment'
    CourseProgress:
      description: Прогресс по курсу
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/CourseProgress'
    CourseProgressList:
      description: Прогресс по курсам
      content:
        application/json:
          schema:
            type: array
            items:
              $ref: '#/components/schemas/CourseProgress'
...
  schemas:
    ID:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required: [error]
      properties:
        error:
          type: object
          additionalProperties: false
          required: [message, status]
          properties:
            message:
              type: string
              minLength: 1
              example: Сотрудник не найден
            status:
              type: integer
              minimum: 400
              maximum: 599
              example: 404
...
    AccessSettings:
      type: object
      additionalProperties: false
      required: [scope, departmentIds, positionIds, userIds]
      properties:
        scope:
          $ref: '#/components/schemas/AccessScope'
        departmentIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        positionIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        userIds:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
    ArticleSection:
      type: object
      additionalProperties: false
      required: [id, name, parentId, order, access]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        name:
          type: string
          minLength: 1
        parentId:
          type: [string, 'null']
          format: uuid
          x-go-type: nullable.Nullable[string]
          x-go-type-skip-optional-pointer: true
        order:
          type: integer
          minimum: 0
        access:
          $ref: '#/components/schemas/AccessSettings'
...
    CreateArticleSectionInput:
      type: object
      additionalProperties: false
      required: [name, parentId]
      properties:
        name:
          type: string
          minLength: 1
        parentId:
          type: [string, 'null']
          format: uuid
          x-go-type: nullable.Nullable[string]
          x-go-type-skip-optional-pointer: true
        access:
          $ref: '#/components/schemas/AccessSettings'
...
    CourseStatus:
      type: string
      enum: [draft, published]
    Course:
      type: object
      additionalProperties: false
      required: [id, title, status, authorId, sequential, createdAt, updatedAt]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        title:
          type: string
        description:
          type: string
        coverUrl:
          type: string
          format: uri
        status:
          $ref: '#/components/schemas/CourseStatus'
        authorId:
          $ref: '#/components/schemas/ID'
        sequential:
          type: boolean
        deadlineDays:
          type: integer
          minimum: 1
        createdAt:
          $ref: '#/components/schemas/ISODateTime'
        updatedAt:
          $ref: '#/components/schemas/ISODateTime'
...
    CreateCourseInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          $ref: '#/components/schemas/CourseStatus'
        sequential:
          type: boolean
        deadlineDays:
          type: integer
          minimum: 1
    CreateCourseFromKbInput:
      type: object
      additionalProperties: false
      required: [title, mode, sectionIds, articleIds]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        sequential:
          type: boolean
        deadlineDays:
          type: integer
          minimum: 1
        mode:
          $ref: '#/components/schemas/LessonSourceMode'
        sectionIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
        articleIds:
          type: array
          minItems: 1
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ID'
    UpdateCourseInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/CourseStatus'
        sequential:
          type: boolean
        deadlineDays:
          type: integer
          minimum: 0
          description: Значение 0 очищает дедлайн, как в текущем frontend-контракте.
    CreateCourseSectionInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
    UpdateCourseSectionInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
    CreateLessonInput:
      type: object
      additionalProperties: false
      required: [courseId, sectionId, title]
      properties:
        courseId:
          $ref: '#/components/schemas/ID'
        sectionId:
          $ref: '#/components/schemas/ID'
        title:
          type: string
          minLength: 1
        content:
          $ref: '#/components/schemas/RichTextContent'
        sourceArticleId:
          $ref: '#/components/schemas/ID'
        sourceMode:
          $ref: '#/components/schemas/LessonSourceMode'
    UpdateLessonInput:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        title:
          type: string
        content:
          $ref: '#/components/schemas/RichTextContent'
        sourceArticleId:
          $ref: '#/components/schemas/ID'
        sourceMode:
          $ref: '#/components/schemas/LessonSourceMode'
    MoveLessonInput:
      type: object
      additionalProperties: false
      required: [sectionId, order]
      properties:
        sectionId:
          $ref: '#/components/schemas/ID'
        order:
          type: integer
          minimum: 0
    UpsertQuizInput:
      type: object
      additionalProperties: false
      required: [lessonId, questions, passingScore]
      properties:
        id:
          $ref: '#/components/schemas/ID'
        lessonId:
          $ref: '#/components/schemas/ID'
        questions:
          type: array
          items:
            $ref: '#/components/schemas/QuizQuestion'
        passingScore:
          type: integer
          minimum: 0
          maximum: 100
        maxAttempts:
          type: integer
          minimum: 1
    AssignCourseInput:
      type: object
      additionalProperties: false
      required: [courseId, assigneeType]
      properties:
        courseId:
          $ref: '#/components/schemas/ID'
        assigneeType:
          $ref: '#/components/schemas/AssigneeType'
        assigneeId:
          $ref: '#/components/schemas/ID'
        dueDate:
          $ref: '#/components/schemas/ClearableDateTime'
    MarkLessonCompleteInput:
      type: object
      additionalProperties: false
      required: [courseId]
      properties:
        courseId:
          $ref: '#/components/schemas/ID'
        userId:
          $ref: '#/components/schemas/ID'
...